Cleaned up when runtime dependencies are checked for, added dep check for dmenu itself, more formatting
