+++ src/config/src/config.rs
@@ -1,11 +1,9 @@
 use std::env;
 use std::fs::File;
 use std::io::{Read, Write};
-use std::process::Command;
 use man_dmenu::*;
 use std::path::PathBuf;
 use itertools::Itertools;
-use termcolor::{Color, ColorChoice, ColorSpec, StandardStream, WriteColor};
 use yaml_rust::{YamlEmitter, Yaml, yaml};
 
 mod util;
@@ -14,8 +12,15 @@ use crate::util::*;
 fn main() {
     let target_path = PathBuf::from(env!("BUILD_TARGET_PATH"));
     let build_path = PathBuf::from(env!("BUILD_PATH"));
-    let mut stdout = StandardStream::stdout(ColorChoice::Always);
+    let mut build_failed = false;
+
+    // Check for dependencies
+    if run_build_command(&format!("sh checkdeps.sh"), &"src/",
+			 &format!("dependency check")) {
+	build_failed = true;
+    }
     
+    // On to plugins
     // First, figure out what plugins we are using
     let plugins = get_selected_plugin_list();
 
@@ -33,7 +38,6 @@ fn main() {
 
     // For every plugin, check if it has arguements. If so, add them to clap and overrider
     // While we're here, set proc_use to watch the plugin entry points
-    let mut failed = false;
     for plugin in plugins {
 	let mut plugin_yaml = get_yaml(&format!("../plugins/{}/plugin.yml", plugin));
 	
@@ -58,43 +62,13 @@ fn main() {
 	}
 
 	if let Some(build_command) = get_yaml_top_level(&mut plugin_yaml, "build") {
-	    let mut command = Command::new("sh");
-	    let dir = format!("../plugins/{}/", plugin);
-	    command.current_dir(dir);
-	    let output = command.arg("-c")
-		.arg(build_command).output()
-		.expect("failed to execute plugin build command");
-	    let stdout_cmd = String::from_utf8_lossy(&output.stdout);
-	    let stderr_cmd = String::from_utf8_lossy(&output.stderr);
-	    let stdout_ref = stdout_cmd.trim_end();
-	    let stderr_ref = stderr_cmd.trim_end();
-	    if stdout_ref.len() > 0 {
-		println!("{}", stdout_ref);
+	    if run_build_command(build_command, &format!("../plugins/{}/", plugin),
+				 &format!("plugin {}", plugin)) {
+		build_failed = true;
 	    }
-	    if stderr_ref.len() > 0 {
-		stdout.set_color(ColorSpec::new().set_fg(Some(Color::Red)))
-		    .expect("Could not grab stdout!");
-		println!("{}", stderr_ref);
-	    }
-	    if output.status.success() {
-		stdout.set_color(ColorSpec::new().set_fg(Some(Color::Green)).set_bold(true))
-		    .expect("Could not grab stdout!");
-		print!("PASS");
-	    } else {
-		stdout.set_color(ColorSpec::new().set_fg(Some(Color::Red)).set_bold(true))
-		    .expect("Could not grab stdout!");
-		print!("FAIL");
-		failed = true;
-	    }
-	    stdout.set_color(ColorSpec::new().set_bold(true))
-		.expect("Could not grab stdout!");
-	    print!(" Running build command for plugin {}", plugin);
-	    stdout.set_color(&ColorSpec::new())
-		    .expect("Could not grab stdout!");
-	    println!(""); // make sure colors are flushed
 	}
     }
-    if failed {
+    if build_failed {
 	std::process::exit(1);
     }
